fix crash when opening about view (#229)

+++ shared/ui/src/commonMain/kotlin/com/androidmakers/ui/about/AboutScreen.kt
@@ -57,7 +57,7 @@ fun AboutScreen(
         text = stringResource(
             MR.strings.version,
             versionName,
-            versionCode
+            versionCode.toInt()
         ),
     )
   }
